requirements.txt
Flask==2.2.2
google-generativeai==1.0.0
python-dotenv==0.21.1
requests==2.28.2
